Move source files into a Python package
Relative imports are now specified as such.
