+++ pyproject.toml
@@ -1,0 +1,21 @@
+[build-system]
+requires = [ "flit_core >=3.2,<4" ]
+build-backend = "flit_core.buildapi"
+
+[project]
+name = "pacfix"
+description = "PAC-learning-based program systhesizer"
+readme = "README.md"
+requires-python = ">=3.9"
+authors = [ { name = "한승헌", email = "shhan@unist.ac.kr" } ]
+maintainers = [ { name = "한승헌", email = "shhan@unist.ac.kr" } ]
+keywords = [ "apr" ]
+classifiers = [
+    "Development Status :: 1 - Planning",
+    "Environment :: Console",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python" ]
+dynamic = [ "version" ]
+dependencies = [ "pysmt" ]
+urls = { GitHub = "https://github.com/hsh814/pacfix-python" }
+scripts = { pacfix = "pacfix.__main__:main" }
