+++ src/pacfix/invariant.py
